testing/web-platform/tests/portals/portals-host-post-message.sub.html

<!DOCTYPE html>
<title>Test postMessage on PortalHost</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <script>
    function createPortal(portalSrc) {
      var portal = document.createElement("portal");
      portal.src = portalSrc;
      return new Promise((resolve, reject) => {
        portal.addEventListener("message", e => {
          if (e.data == "loaded")
            resolve(portal);
        });
        document.body.appendChild(portal);
      });
    }

    async function createPortalAndLoopMessage(portalSrc, params) {
      var portal = await createPortal(portalSrc);
      var waitForResponse = new Promise((resolve, reject) => {
        portal.addEventListener("message", e => { resolve(e); });
      });
      portal.postMessage(params, "*");
      return waitForResponse;
    }

    const crossOriginUrl = "http://{{hosts[alt][www]}}:{{ports[http][0]}}/portals/resources/portal-host-post-message.sub.html";

    promise_test(async () => {
      var {data, origin} = await createPortalAndLoopMessage(
          "resources/portal-host-post-message.sub.html", ["test", "*"]);
      assert_equals(data, "test");
      assert_equals(origin, "http://{{host}}:{{ports[http][0]}}");
    }, "Message received after postMessage from portal host");

    promise_test(async () => {
      var {data, origin} = await createPortalAndLoopMessage(
          crossOriginUrl, ["test", "*"]);
      assert_equals(data, "test");
      assert_equals(origin, "http://{{hosts[alt][www]}}:{{ports[http][0]}}");
    }, "Message received after postMessage from portal host in cross-origin-portal");

    promise_test(async () => {
      var {data, origin} = await createPortalAndLoopMessage(
          "resources/portal-host-post-message.sub.html", ["test"]);
      assert_equals(data, "test");
      assert_equals(origin, "http://{{host}}:{{ports[http][0]}}");
    }, "Message received from same-origin portal host with no target origin specified");

    promise_test(async () => {
      var {data, origin} = await createPortalAndLoopMessage(
         crossOriginUrl, ["test", "http://{{host}}:{{ports[http][0]}}"]);
      assert_equals(data, "test");
    }, "Message received from cross-origin portal host with target origin correctly specified");

    promise_test(async () => {
       var receiveMessage = new Promise((resolve, reject) => {
         var bc = new BroadcastChannel("portal-host-post-message-after-activate");
         bc.onmessage = e => { resolve(e); };
       });
       const portalUrl = encodeURIComponent(
          "portal-host-post-message-after-activate.html");
       window.open(`resources/portal-embed-and-activate.html?url=${portalUrl}`);
       var message = await receiveMessage;
       assert_equals(message.data, "InvalidStateError");
    }, "Calling postMessage after receiving onactivate event should fail");

    promise_test(() => {
      var portal = document.createElement("portal");
      portal.src = "resources/portal-host-post-message-navigate-1.html";
      var count = 0;
      var waitForMessages = new Promise((resolve, reject) => {
        portal.addEventListener("message", e => {
          count++;
          if (count == 2)
            resolve();
        });
      });
      document.body.appendChild(portal);
      return waitForMessages;
    }, "postMessage before and after portal navigation should work");

    async_test(t => {
      createPortalAndLoopMessage("resources/portal-host-post-message.sub.html",
          ["test", "http://{{hosts[alt][www]}}:{{ports[http][0]}}"])
          .then(t.step_func(() => { assert_unreached("message delivered"); }));
      t.step_timeout(t.done, 2000);
    }, "Message should not be received from portal host with target set to different origin");

    async_test(t => {
      createPortalAndLoopMessage(crossOriginUrl, ["test"]).then(t.step_func(() => {
        assert_unreached("message delivered");
      }));
      t.step_timeout(t.done, 2000);
    }, "Message should not be received cross origin-portal host with no target origin set");
  </script>
</body>
